Update rom original

+++ Tools/sitl_run.sh
@@ -14,6 +14,10 @@ echo program: $program
 echo model: $model
 echo build_path: $build_path
 
+mkdir -p $build_path/src/firmware/posix/rootfs/fs/microsd
+mkdir -p $build_path/src/firmware/posix/rootfs/eeprom
+touch $build_path/src/firmware/posix/rootfs/eeprom/parameters
+
 if [ "$chroot" == "1" ]
 then
 	chroot_enabled=-c
@@ -29,7 +33,7 @@ then
 	model="iris"
 fi
 
-if [ "$#" != 5 ]
+if [ "$#" -lt 5 ]
 then
 	echo usage: sitl_run.sh rc_script debugger program model build_path
 	echo ""
@@ -48,50 +52,59 @@ fi
 
 set -e
 
+cd $build_path/..
 cp Tools/posix_lldbinit $build_path/src/firmware/posix/.lldbinit
 cp Tools/posix.gdbinit $build_path/src/firmware/posix/.gdbinit
 
 SIM_PID=0
 
-if [ "$program" == "jmavsim" ] && [ "$no_sim" == "" ]
+if [ "$program" == "jmavsim" ] && [ ! -n "$no_sim" ]
 then
 	cd Tools/jMAVSim
-	ant
-	java -Djava.ext.dirs= -cp lib/*:out/production/jmavsim.jar me.drton.jmavsim.Simulator -udp 127.0.0.1:14560 &
+	ant create_run_jar copy_res
+	cd out/production
+	java -Djava.ext.dirs= -jar jmavsim_run.jar -udp 127.0.0.1:14560 &
 	SIM_PID=`echo $!`
-elif [ "$program" == "gazebo" ] && [ "$no_sim" == "" ]
+	cd ../..
+elif [ "$program" == "gazebo" ] && [ ! -n "$no_sim" ]
 then
 	if [ -x "$(command -v gazebo)" ]
 	then
 		# Set the plugin path so Gazebo finds our model and sim
-		export GAZEBO_PLUGIN_PATH=$curr_dir/Tools/sitl_gazebo/Build:${GAZEBO_PLUGIN_PATH}
+		export GAZEBO_PLUGIN_PATH=$curr_dir/build_gazebo:${GAZEBO_PLUGIN_PATH}
 		# Set the model path so Gazebo finds the airframes
 		export GAZEBO_MODEL_PATH=${GAZEBO_MODEL_PATH}:$curr_dir/Tools/sitl_gazebo/models
 		# The next line would disable online model lookup, can be commented in, in case of unstable behaviour.
 		# export GAZEBO_MODEL_DATABASE_URI=""
 		export SITL_GAZEBO_PATH=$curr_dir/Tools/sitl_gazebo
-		mkdir -p Tools/sitl_gazebo/Build
-		cd Tools/sitl_gazebo/Build
-		cmake -Wno-dev ..
-		make -j4
-		gzserver --verbose ../worlds/${model}.world &
+		make --no-print-directory gazebo_build
+
+		gzserver --verbose $curr_dir/Tools/sitl_gazebo/worlds/${model}.world &
 		SIM_PID=`echo $!`
-		gzclient --verbose &
-		GUI_PID=`echo $!`
+
+		if [[ -n "$HEADLESS" ]]; then
+			echo "not running gazebo gui"
+		else
+			gzclient --verbose &
+			GUI_PID=`echo $!`
+		fi
 	else
 		echo "You need to have gazebo simulator installed!"
 		exit 1
 	fi
-elif [ "$program" == "replay" ] && [ "$no_sim" == "" ]
+elif [ "$program" == "replay" ] && [ ! -n "$no_sim" ]
 then
 	echo "Replaying logfile: $logfile"
 	# This is not a simulator, but a log file to replay
+
+	# Check if we need to creat a param file to allow user to change parameters
+	if ! [ -f "${build_path}/src/firmware/posix/rootfs/replay_params.txt" ]
+		then
+		touch ${build_path}/src/firmware/posix/rootfs/replay_params.txt
+	fi
 fi
 
 cd $build_path/src/firmware/posix
-mkdir -p rootfs/fs/microsd
-mkdir -p rootfs/eeprom
-touch rootfs/eeprom/parameters
 
 if [ "$logfile" != "" ]
 then
@@ -124,5 +137,7 @@ then
 elif [ "$program" == "gazebo" ]
 then
 	kill -9 $SIM_PID
-	kill -9 $GUI_PID
+	if [[ ! -n "$HEADLESS" ]]; then
+		kill -9 $GUI_PID
+	fi
 fi
